Kit::Log - Move very first version to its own gem
