+++ domains/kit-auth/Gemfile
@@ -14,6 +14,7 @@ gem 'kit-doc',          path: '../../libraries/kit-doc' # DEV MODE
 gem 'kit-domain',       path: '../../libraries/kit-domain' # DEV MODE
 gem 'kit-engine',       path: '../../libraries/kit-engine' # DEV MODE
 gem 'kit-error',        path: '../../libraries/kit-error' # DEV MODE
+gem 'kit-log',          path: '../../libraries/kit-log' # DEV MODE
 gem 'kit-organizer',    path: '../../libraries/kit-organizer' # DEV MODE
 gem 'kit-pagination',   path: '../../libraries/kit-pagination' # DEV MODE
 gem 'kit-router',       path: '../../libraries/kit-router' # DEV MODE
