+++ libraries/kit-log/.rubocop.yml
@@ -1,0 +1,1 @@
+inherit_from: ../../.rubocop_shared.yml
