+++ libraries/kit-log/Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{ repo }.git" }
+
+gemspec
