+++ libraries/kit-log/README.md
@@ -1,0 +1,20 @@
+<!--pp {} -->
+<img align="left" width="50" height="90" src="https://raw.githubusercontent.com/rubykit/kit/master/docs/assets/images/rubykit-framework-logo.svg">
+<!-- pp-->
+
+[Kit::Log]: https://github.com/rubykit/kit/tree/master/libraries/kit-log
+
+# Kit::Log
+
+[Kit::Log] is the logging library for Kit.
+
+## Usage
+
+1) Add `kit-log` to your Gemfile
+2) Run: `bundle install`
+
+## Copyright & License
+
+Copyright (c) 2020, Nathan Appere.
+
+[Kit::Log] is licensed under [MIT License](MIT_LICENSE.md).
