+++ libraries/kit-log/lib/kit-log.rb
@@ -1,0 +1,4 @@
+# rubocop:disable Naming/FileName
+# rubocop:enable Naming/FileName
+
+require_relative 'kit/log'
