+++ libraries/kit-log/lib/kit/log/backends.rb
@@ -1,0 +1,4 @@
+# Namespace for all output backends
+module Kit::Log::Backends
+
+end
