+++ libraries/kit-log/lib/kit/log/version.rb
@@ -1,0 +1,8 @@
+module Kit # rubocop:disable Style/Documentation
+end
+
+module Kit::Log
+
+  VERSION = '0.1.0'
+
+end
